refactor: Filter 훅을 유틸함수로 변경
